Add files via upload
multi-modal (text + tables + images) RAG, Version_1, by following articles_3 and _4:
# (https://github.com/langchain-ai/langchain/blob/master/cookbook/Semi_Structured_RAG.ipynb)
# (https://github.com/langchain-ai/langchain/blob/master/cookbook/Semi_structured_and_multi_modal_RAG.ipynb) 
use Google Gemini (gemini-1.5-flash) instead of OpenAI because of API limit
BUT still it can not work on a 11-page PDF (AKAP1.pdf), always over the limit, even after adjustment. Table summary was okay, but text summary is not finished. So that it can not go on.